add base page that other pages can inherit from
